+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Feature
 - Refactored the way to translate a story: no cloned stories, everything is handled with translation key with yaml files.
+- Confirm screen when clicking on "new game" if a save is detected
 
 ### Fixes
 - Main music in main screen stopping when going in options - 26.2